The list of similar shows returned by the API is also a paged list that returns the exact same data types as the top rated ones so the way to retrieve these paged lists has been generalised a bit and then made concrete only on the specific calls that need slightly different input.
The details screen now displays a horizontally scrollable list of related shows. Tapping on one of those shows takes the user to the details screen for that show.

The show homepage url that the user can click to get more information is also added to the bottom of the details screen.